# scrivia-v2-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scrivia API v2 - LLM Tools",
    "description": "API Scrivia v2 pour l'intégration LLM avec support complet des notes, dossiers et classeurs. Tous les endpoints supportent les slugs et UUIDs.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://scrivia.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "source_title": { "type": "string", "maxLength": 255 },
          "markdown_content": { "type": "string" },
          "html_content": { "type": "string" },
          "header_image": { "type": "string", "format": "uri" },
          "folder_id": { "type": "string", "format": "uuid" },
          "classeur_id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "description": { "type": "string", "maxLength": 500 },
          "visibility": { "type": "string", "enum": ["private", "shared", "members", "public"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "source_title", "user_id", "created_at", "updated_at"]
      },
...
      "AddContentPayload": {
        "type": "object",
        "properties": {
          "content": { "type": "string", "minLength": 1 }
        },
        "required": ["content"]
      },
...
      "Folder": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "maxLength": 255 },
          "parent_id": { "type": "string", "format": "uuid", "nullable": true },
          "classeur_id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "classeur_id", "user_id", "created_at", "updated_at"]
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "object" }
        },
        "required": ["success"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "integer" }
        },
        "required": ["error"]
      }
    }
  },
  "paths": {
...
    "/api/v2/note/{ref}/content": {
      "get": {
        "operationId": "getNoteContent",
        "summary": "Récupérer le contenu d'une note",
        "description": "Récupérer le contenu markdown et HTML d'une note",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "responses": {
          "200": {
            "description": "Contenu récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "markdown_content": { "type": "string" },
                    "html_content": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
    "/api/v2/note/{ref}/add-content": {
      "post": {
        "operationId": "addContentToNote",
        "summary": "Ajouter du contenu à une note",
        "description": "Ajouter du contenu markdown à la fin d'une note existante",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddContentPayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contenu ajouté avec succès",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Note" }
              }
            }
          },
          "400": { "$ref": "#/components/schemas/ErrorResponse" },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
    "/api/v2/folder/{ref}/delete": {
      "delete": {
        "operationId": "deleteFolder",
        "summary": "Supprimer un dossier",
        "description": "Supprimer définitivement un dossier et tout son contenu",
        "tags": ["Dossiers"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug du dossier"
          }
        ],
        "responses": {
          "200": {
            "description": "Dossier supprimé avec succès",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
    "/api/v2/folder/{ref}/tree": {
      "get": {
        "operationId": "getFolderTree",
        "summary": "Récupérer l'arborescence d'un dossier",
        "description": "Récupérer la structure complète d'un dossier avec ses sous-dossiers et notes",
        "tags": ["Dossiers"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug du dossier"
          }
        ],
        "responses": {
          "200": {
            "description": "Arborescence récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "folder": { "$ref": "#/components/schemas/Folder" },
                    "children": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
  }
} 