# src/app/api/v2/note/[ref]/content/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { createClient } from '@supabase/supabase-js';
import { logApi } from '@/utils/logger';
import { V2ResourceResolver } from '@/utils/v2ResourceResolver';
import { getAuthenticatedUser } from '@/utils/authUtils';

// 🔧 CORRECTIONS APPLIQUÉES:
// - Authentification simplifiée via getAuthenticatedUser uniquement
// - Suppression de la double vérification d'authentification
// - Client Supabase standard sans token manuel
// - Plus de 401 causés par des conflits d'authentification

const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!;
const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!;

export async function GET(
  request: NextRequest,
  { params }: { params: Promise<{ ref: string }> }
): Promise<NextResponse> {
  const startTime = Date.now();
  const { ref } = await params;
  const clientType = request.headers.get('X-Client-Type') || 'unknown';
  const context = {
    operation: 'v2_note_content',
    component: 'API_V2',
    ref,
    clientType
  };

  logApi.info(`🚀 Début récupération contenu note v2 ${ref}`, context);

  // 🔐 Authentification simplifiée
  const authResult = await getAuthenticatedUser(request);
  if (!authResult.success) {
    logApi.info(`❌ Authentification échouée: ${authResult.error}`, context);
    return NextResponse.json(
      { error: authResult.error },
      { status: authResult.status || 401, headers: { "Content-Type": "application/json" } }
    );
  }

  const userId = authResult.userId!;
  
  // 🔧 CORRECTION: Client Supabase standard, getAuthenticatedUser a déjà validé
  const supabase = createClient(supabaseUrl, supabaseAnonKey);

  // Résoudre la référence (UUID ou slug)
  const resolveResult = await V2ResourceResolver.resolveRef(ref, 'note', userId, context);
  if (!resolveResult.success) {
    return NextResponse.json(
      { error: resolveResult.error },
      { status: resolveResult.status, headers: { "Content-Type": "application/json" } }
    );
  }

  const noteId = resolveResult.id;

  // 🔐 Vérification des permissions simplifiée
  try {
    // Vérifier directement si l'utilisateur a accès à cette note
    const { data: article, error: articleError } = await supabase
      .from('articles')
      .select('user_id, share_settings')
      .eq('id', noteId)
      .single();
    
    if (articleError || !article) {
      logApi.info(`❌ Note non trouvée: ${noteId}`, context);
      return NextResponse.json(
        { error: 'Note non trouvée' },
        { status: 404, headers: { "Content-Type": "application/json" } }
      );
    }

    // ✅ ACCÈS AUTORISÉ si :
    // 1. L'utilisateur est le propriétaire de la note
    // 2. OU la note est accessible via lien (link-private, link-public, limited, scrivia)
    const isOwner = article.user_id === userId;
    const isAccessible = article.share_settings?.visibility !== 'private';
    
    if (!isOwner && !isAccessible) {
      logApi.info(`❌ Accès refusé pour note ${noteId}`, context);
      return NextResponse.json(
        { error: 'Accès refusé' },
        { status: 403, headers: { "Content-Type": "application/json" } }
      );
    }

    // Récupérer le contenu complet de la note
    const { data: note, error: noteError } = await supabase
      .from('articles')
      .select('id, source_title, markdown_content, html_content, header_image, created_at, updated_at')
      .eq('id', noteId)
      .single();

    if (noteError || !note) {
      logApi.info(`❌ Erreur récupération contenu: ${noteError?.message}`, context);
      return NextResponse.json(
        { error: 'Erreur lors de la récupération du contenu' },
        { status: 500, headers: { "Content-Type": "application/json" } }
      );
    }

    const apiTime = Date.now() - startTime;
    logApi.info(`✅ Contenu note v2 récupéré en ${apiTime}ms`, context);

    return NextResponse.json({
      success: true,
      note: {
        id: note.id,
        title: note.source_title,
        markdown_content: note.markdown_content,
        html_content: note.html_content,
        header_image: note.header_image,
        created_at: note.created_at,
        updated_at: note.updated_at
      }
    }, { headers: { "Content-Type": "application/json" } });

  } catch (error) {
    const errorMessage = error instanceof Error ? error.message : 'Erreur inconnue';
    logApi.error(`❌ Erreur inattendue: ${errorMessage}`, context);
    
    return NextResponse.json(
      { error: 'Erreur interne du serveur' },
      { status: 500, headers: { "Content-Type": "application/json" } }
    );
  }
} 